Rename module back to `vrt.py`
