+++ yt_dlp/extractor/vrt.py
